Exit _build.sh with error when any of it's subtasks fail (#34)
`set -ev` prints every command executed and fails the entire script when
any command fails. This prevents false-positive cases where only one of
the tasks fails. For instance when there are LaTeX specific errors which
only cause the `bookdown::pdf_book` script to fail. In this case Travis still
marks a pull request as if the build passes.